requirements.txt
flask
flask-cors
transformers
torch
torchvision
pillow
numpy
scipy
